Move api route path
