+++ app/api/suppliers/[handle]/check-availability/route.ts
